OnButtonDownClick ball launches
On button click the balls velocity is changed to give the illusion the ball is pushed upwards. This can only be pressed once.

Fixed bug: The LocktoPaddle method was being updated every frame so the ball wouldn't move, had to implement a boolean to determine if the ball has been launched then stop locking the ball.